3-code/cli/pyproject.toml
[project]
name = "vision-cli"
version = "0.0.1"
description = "Operator CLI for project-agent-system."
requires-python = "==3.12.*"
dependencies = [
    "typer>=0.15,<0.16",
    "httpx>=0.27,<0.28",
    "pydantic>=2.9,<3",
    "python-dotenv>=1.0,<2",
    "rich>=13.9,<14",
]

[project.scripts]
vision = "app.main:cli"

[dependency-groups]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<0.25",
    "ruff>=0.7,<0.8",
    "mypy>=1.13,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "UP",
    "SIM",
    "TCH",
    "RUF",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
